Moved notifications into the same file
